+++ frontend/src/pages/Projects/Project/ProjectDetails/BudgetSection/index.tsx
@@ -5,7 +5,7 @@ import { useFormControls } from "hooks";
 import { useFormData } from "hooks/useFormData";
 import { IFormControls } from "types";
 import { tableConfig } from "./tableConfig";
-import { FormConfig } from "./formConfig";
+import { FormConfig } from "./FormConfig";
 
 export const BudgetSection = () => {
   const { projectId } = useParams();